Milestone 1 Code Commit
These files are part of Milestone 1 of the hospital patient manager project. In this Milestone, the Time module (which holds functions to retrieve the current time in GMT - 5:00) and Utils module (which holds many reusable validation and conversion functions to help with the rest of the project) were completed and a Tester program was utilized to help test the functionality of my code. This Module was done between July 1st and July 7th of 2021.