Milestone 2 Code Commit
These files are part of Milestone 2 of the hospital patient manager project. In this Milestone, the Menu module (which holds the necessary class and function for a menu system) and IOAble module (which holds multiple pure virtual functions and multiple helper functions which will be used by other modules in the project for input and output) were added and completed. Furthermore, a new Tester program was also utilized to help test the functionality of my code. This Module was done between July 8th and July 14th of 2021.