Milestone 3 Code Commit
These files are part of Milestone 3 of the hospital patient manager project. In this Milestone, the Ticket module (which holds the class and methods responsible for handling patients' ticket numbers) and Patient module (which holds the class and methods responsible for handling patients' medical and personal information) were added and completed. Minor changes were also made to improve the functionality of the code in the Utils module, as well. A new Tester program was also utilized, and there is the addition of testData.csv and testDataOut.csv files which all work together to help test the functionality of my code. This Module was done between July 15th and July 23rd of 2021.