Milestone 4 Code Commit
These files are part of Milestone 4 of the hospital patient manager/system project. In this Milestone, the TriagePatient module (which holds the class that is responsible for keeping a record of triage patients' information and is inherited from the Patient class in the Patient module) and CovidPatient module (which holds the class that is responsible for keeping a record of COVID patients' information and is also inherited from the Patient class in the Patient module) were added and completed. A new Tester program was also utilized and the previously used testData.csv and testDataOut.csv files are no longer required for testing. This Module was done between July 24th and July 30rd of 2021.