Milestone 5 Code Commit
These files are part of Milestone 5, the last milestone, of the hospital patient system project. In this Milestone, the PreTriage module (responsible for getting, storing, and providing information for pre-triage/pre-admittance patients) was added and completed. Furthermore, due to this being the final milestone of the project, the code throughout the modules was cleaned up and comments were added to help future developers in updating the hospital patient system. A finalized Tester program was also utilized and a few data files were all collectively used to test the finalized version of the hospital patient system. This Module was done between July 31st and August 4th of 2021.